+++ src/newtab/Newtab.tsx
@@ -2,14 +2,14 @@ import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 
-import Grid from './containers/Grid'
+import { Newtab } from './components/Newtab'
 import store from './store/newtab'
 
 declare let module: any
 
 ReactDOM.render(
   <Provider store={store}>
-    <Grid />
+    <Newtab />
   </Provider>,
   document.getElementById('root')
 )