+++ src/newtab/components/Newtab.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+import Grid from '../containers/Grid'
+import Sidebar from '../containers/Sidebar'
+import './Newtab.scss'
+
+export class Newtab extends React.Component {
+  render() {
+    return (
+      <div className={'newtab'}>
+        <Grid />
+        <Sidebar />
+      </div>
+    )
+  }
+}