+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added tile display.
+- Added ability to add tiles.
 - Added redux mock store.
 
 ### Changed